refactor(autoapi): move nested prefix helper into routes builder
