+++ pkgs/standards/autoapi/autoapi/v2/impl/routes_builder.py
@@ -11,7 +11,7 @@ import functools
 import inspect
 import re
 from types import SimpleNamespace
-from typing import Annotated, Any, List, Optional
+from typing import Annotated, Any, List, Optional, Type
 
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import Session
@@ -81,6 +81,21 @@ def _resource_pascal(tab_or_cls: str) -> str:
         if tab_or_cls.islower()
         else tab_or_cls
     )
+
+
+def _nested_prefix(self, model: Type) -> Optional[str]:
+    """Return the user-supplied hierarchical prefix or *None*.
+
+    • If the SQLAlchemy model defines `__autoapi_nested_paths__`
+      → call it and return the result.
+    • Else, fall back to legacy `_nested_path` string if present.
+    • Otherwise → signal ``no nested route wanted`` with ``None``.
+    """
+
+    cb = getattr(model, "__autoapi_nested_paths__", None)
+    if callable(cb):
+        return cb()
+    return getattr(model, "_nested_path", None)
 
 
 # ──────────────────────────────────────────────────────────────────────────────
